refactor(tidy): break #1 - phases 1-4 physical move
Structural moves only. Imports across repo still reference old paths
and will not resolve until recovery. See recovery task at
teams/tasks/backend-tidy/recovery-task-break-1.md.

Phase 1: promote backend/protocols/ -> protocols/ (top-level)
Phase 2: consolidate 12 sandbox-shaped files -> backend/sandboxes/
Phase 3: consolidate 10 auth/avatar/profile files -> backend/identity/
Phase 4: consolidate 6 resource files -> backend/sandboxes/resources/

Preserved compat shell: backend/web/core/supabase_factory.py (external
env files reference this path via LEON_SUPABASE_CLIENT_FACTORY).

32 renames + 6 new __init__.py + 1 compat shell update.
No logic changes.
