--- backend/protocols/__init__.py
@@ -1,1 +1,0 @@
-"""Transport protocol contracts shared across backend boundaries."""
