--- backend/protocols/runtime_read.py
@@ -1,5 +1,0 @@
-"""Compatibility export surface for top-level runtime-read protocols."""
-
-from protocols.runtime_read import AgentThreadActivity, RuntimeThreadActivityReader
-
-__all__ = ["AgentThreadActivity", "RuntimeThreadActivityReader"]
